+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Fix: Include normalization in data dimension name
 - Fix: Allow rendering a single axis instead of enforcing either none or both axis
 - Fix: Rely on pre-normalized data to get bin ID
+- Fix: Connect order
 
 ## v0.16.0